+++ tests/src/jmap/calendar/mod.rs
@@ -9,4 +9,5 @@ pub mod alarm;
 pub mod calendars;
 pub mod event;
 pub mod identity;
+pub mod instance;
 pub mod notification;